+++ extensions/positron-python/package.json
@@ -437,13 +437,6 @@
             },
             {
                 "category": "Python",
-                "command": "python.runAndDebugCell",
-                "icon": "$(debug-alt-small)",
-                "title": "Debug Cell",
-                "enablement": "notebookKernel =~ /^positron.runtime-notebook-kernels\\//"
-            },
-            {
-                "category": "Python",
                 "command": "python.debugDashInTerminal",
                 "icon": "$(debug-alt)",
                 "title": "%python.command.python.debugDashInTerminal.title%",
@@ -1373,10 +1366,6 @@
                 },
                 "when": "!virtualWorkspace && shellExecutionSupported",
                 "hiddenWhen": "true"
-            },
-            {
-                "type": "runtimeNotebook",
-                "label": "Runtime Notebook"
             }
         ],
         "grammars": [
@@ -1852,11 +1841,6 @@
                     "command": "python.createNewFile",
                     "group": "file",
                     "when": "!virtualWorkspace"
-                }
-            ],
-            "notebook/cell/execute": [
-                {
-                    "command": "python.runAndDebugCell"
                 }
             ],
             "view/title": [
