+++ extensions/positron-python/src/client/positron/extension.ts
@@ -16,7 +16,6 @@ import { activateWebAppCommands } from './webAppCommands';
 import { activateWalkthroughCommands } from './walkthroughCommands';
 import { printInterpreterDebugInfo } from './interpreterSettings';
 import { registerLanguageServerManager } from './languageServerManager';
-import { activateRuntimeNotebookDebugging } from './debug';
 
 export async function activatePositron(serviceContainer: IServiceContainer): Promise<void> {
     try {
@@ -91,9 +90,6 @@ export async function activatePositron(serviceContainer: IServiceContainer): Pro
 
         // Activate web application commands.
         activateWebAppCommands(serviceContainer, disposables);
-
-        // Activate debugging for notebooks backed by language runtimes.
-        activateRuntimeNotebookDebugging(disposables);
 
         // Activate walkthrough commands.
         activateWalkthroughCommands(disposables);
