+++ extensions/positron-runtime-debug/package.nls.json
@@ -1,0 +1,5 @@
+{
+	"displayName": "Runtime Debugging",
+	"description": "Debugging support for Positron language runtimes.",
+	"category.notebook": "Notebook"
+}
