+++ extensions/positron-runtime-debug/tsconfig.json
@@ -1,0 +1,19 @@
+{
+	"compilerOptions": {
+		"module": "commonjs",
+		"target": "ES2020",
+		"outDir": "out",
+		"lib": ["ES2020"],
+		"sourceMap": true,
+		"rootDir": "src",
+		"typeRoots": ["./node_modules/@types"],
+		"paths": {
+			"*": ["./node_modules/*"],
+		},
+	},
+	"include": [
+		"src/**/*",
+		"../../src/vscode-dts/vscode.d.ts",
+		"../../src/positron-dts/positron.d.ts",
+	],
+}
